+++ GeneralsMD/Code/GameEngine/Include/GameClient/GadgetPushButton.h
@@ -84,6 +84,7 @@ void GadgetButtonDrawOverlayImage( GameWindow *g, const Image *image );
 void GadgetButtonSetBorder( GameWindow *g, Color color, Bool drawBorder = TRUE );
 void GadgetButtonSetCount( GameWindow *g, Int count );
 void GadgetButtonSetSeconds( GameWindow *g, Int seconds );
+void GadgetButtonSetCost( GameWindow *g, Int cost );
 void GadgetButtonSetBar( GameWindow *g, Int percent, Color color );
 void GadgetButtonSetData(GameWindow *g, void *data);
 void *GadgetButtonGetData(GameWindow *g);